requirements.txt
absl-py==2.1.0
aioice==0.10.1
aiortc==1.13.0
anyio==4.8.0
argcomplete==1.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
av==14.4.0
babel==2.17.0
backcall==0.2.0
bcrypt==4.3.0
beautifulsoup4==4.8.0
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.7
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
docopt==0.6.2
docx2txt==0.8
EbookLib==0.17.1
eventlet==0.40.0
executing==2.2.0
extract-msg==0.23.1
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
flatbuffers==25.1.24
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.2.0
gast==0.6.0
git-filter-repo==2.47.0
google-crc32c==1.7.1
google-pasta==0.2.0
greenlet==3.2.2
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
IMAPClient==2.1.0
ipykernel==6.29.5
ipython==8.12.3
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.2
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==6.0.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mysql-connector-python==9.2.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
olefile==0.46
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.23.0
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMuPDF==1.26.0
pyOpenSSL==25.1.0
pyparsing==3.2.1
pypdf==6.1.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-engineio==4.12.1
python-json-logger==3.3.0
python-pptx==0.6.18
python-socketio==5.13.0
pytz==2025.1
pywin32==309
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
simple-websocket==1.1.0
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
SpeechRecognition==3.8.1
srt==3.5.3
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
textract==1.6.3
tinycss2==1.4.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==1.5.1
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.3.0
vosk==0.3.45
waitress==3.0.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
xlrd==1.2.0
xlsxwriter==3.2.9
yarg==0.1.9
